# Conflicts:
#	WooCommerce/Classes/ViewRelated/MainTabBarController.swift
#	WooCommerce/Classes/ViewRelated/Notifications/ReviewsViewModel.swift
#	WooCommerce/Classes/ViewRelated/Orders/Order Details/OrderLoaderViewController.swift
#	Yosemite/Yosemite/Actions/NotificationAction.swift
#	Yosemite/Yosemite/Stores/NotificationStore.swift
#	Yosemite/YosemiteTests/Stores/NotificationStoreTests.swift

+++ Networking/Networking/Mapper/OrderCountMapper.swift
@@ -9,7 +9,7 @@ struct OrderCountMapper: Mapper {
     ///
     /// We're injecting this field via `JSONDecoder.userInfo` because SiteID is not returned in any of the Order Endpoints.
     ///
-    let siteID: Int
+    let siteID: Int64
 
     /// (Attempts) to convert a dictionary into OrderCount.
     ///
